backend/requirements.txt
beanie
fastapi
motor
pika
psycopg2
PyJWT
python-dotenv
regex
requests
SQLAlchemy
typing
urllib3
uvicorn
